# configs/openapi_spec_dredd.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ContractIQ PetStore API",
    "description": "A lightweight PetStore API for intelligent API testing and validation",
    "version": "1.0.0",
    "contact": {
      "name": "ContractIQ"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets": {
      "get": {
        "summary": "List Pets",
        "operationId": "list_pets",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PetStatus"
            }
          },
          {
            "name": "species",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Pet",
        "operationId": "create_pet",
        "tags": [
          "Pets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}": {
      "get": {
        "summary": "Get Pet",
        "operationId": "get_pet",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sample-pet-id-001"
            },
            "example": "sample-pet-id-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Pet details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "404": {
            "description": "Pet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Pet",
        "operationId": "update_pet",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sample-pet-id-001"
            },
            "example": "sample-pet-id-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pet updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "404": {
            "description": "Pet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Pet",
        "operationId": "delete_pet",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sample-pet-id-001"
            },
            "example": "sample-pet-id-001"
          }
        ],
        "responses": {
          "204": {
            "description": "Pet deleted"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "summary": "Create Order",
        "operationId": "create_order",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Pet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Orders",
        "operationId": "list_orders",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "summary": "Get Order",
        "operationId": "get_order",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sample-order-id-001"
            },
            "example": "sample-order-id-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "summary": "Get Stats",
        "operationId": "get_stats",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "API statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_pets": {
                      "type": "integer"
                    },
                    "total_orders": {
                      "type": "integer"
                    },
                    "pets_by_status": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PetStatus": {
        "type": "string",
        "enum": [
          "available",
          "pending",
          "sold"
        ]
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "placed",
          "approved",
          "delivered"
        ]
      },
      "PetCreate": {
        "type": "object",
        "required": [
          "name",
          "species",
          "age",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "Dog"
          },
          "age": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30,
            "example": 3
          },
          "status": {
            "$ref": "#/components/schemas/PetStatus"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "friendly"
            ]
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 10000,
            "example": 250.0
          }
        },
        "example": {
          "name": "Buddy",
          "species": "Dog",
          "age": 3,
          "price": 250.0
        }
      },
      "Pet": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PetCreate"
          },
          {
            "type": "object",
            "required": [
              "id",
              "created_at"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              }
            }
          }
        ]
      },
      "PetUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "Dog"
          },
          "age": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30,
            "example": 3
          },
          "status": {
            "$ref": "#/components/schemas/PetStatus"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "friendly"
            ]
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 10000,
            "example": 250.0
          }
        },
        "example": {
          "name": "Buddy",
          "species": "Dog",
          "age": 3,
          "price": 250.0
        }
      },
      "OrderCreate": {
        "type": "object",
        "required": [
          "pet_id",
          "quantity"
        ],
        "properties": {
          "pet_id": {
            "type": "string",
            "example": "sample-pet-id-001"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 1
          }
        },
        "example": {
          "pet_id": "sample-pet-id-001",
          "quantity": 1
        }
      },
      "Order": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrderCreate"
          },
          {
            "type": "object",
            "required": [
              "id",
              "status",
              "total_price",
              "created_at"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "status": {
                "$ref": "#/components/schemas/OrderStatus"
              },
              "total_price": {
                "type": "number"
              },
              "created_at": {
                "type": "string"
              }
            }
          }
        ]
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "timestamp",
          "version"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      }
    }
  }
}